Update interface.py
I updated the the interface code to look like mine since I was not getting any errors like our testers were. Also, added some error handling for when editing the user income by name so that the user has to enter the correct info or it will not let them change anything.